Removed most of the chat code to prepare for fresh restart.
